Delete duplicate code
